feat(documents): native Effect.gen IDP pipeline hop loop
Stage skip/dry-run/execute(+retry)/Merkle/onHop in Effect.gen like
memory ingest; keep Promise runIdpPipeline façade for tests and callers.
