+++ docs/design/modularization-implementation-status.md
@@ -235,17 +235,18 @@ From enablement §5.4 and the Effect plan §8:
 
 ## 7. Effect-TS migration status
 
-| Area                                                                             | Status                                                                                                                |
-| -------------------------------------------------------------------------------- | --------------------------------------------------------------------------------------------------------------------- |
-| `effect` dependency                                                              | ✅ Pinned in `clawql-api`                                                                                             |
-| `SearchService` / `ExecuteService`                                               | ✅ Live Layers; MCP uses `getClawqlApi().run(Effect…)`                                                                |
-| `AuditLive`                                                                      | ✅ Composed in `createClawQLApi()`                                                                                    |
-| `Plugin` / `PluginRegistry`                                                      | ✅ Effect `register` / `beforeCallTool`                                                                               |
-| Extracted packages (`memory`, `documents`, `automation`, `sandbox`, `ouroboros`) | 🚧 Layer wrappers shipped; memory ingest/recall + ouroboros hot paths use native `Effect.gen` (IO still `tryPromise`) |
-| `@effect/schema` at boundaries                                                   | ❌ Zod remains at MCP tool registration                                                                               |
-| Horizontal `Plugin` Layers                                                       | ✅ All tiers via `composeHorizontalPluginLayers()`                                                                    |
-| Operator dynamic Layer list from CRD                                             | ✅ `composeHorizontalPluginLayersFromTierSpec()` maps `ClawQLHorizontalTierSpec` → Layers                             |
-| `MemoryIngestService` / vault post-sync                                          | ✅ Native `Effect.gen` stages prepare → vault write → `MemoryDbService` sync (no nested `runMemoryEffect`)            |
+| Area                                                                             | Status                                                                                                                                       |
+| -------------------------------------------------------------------------------- | -------------------------------------------------------------------------------------------------------------------------------------------- |
+| `effect` dependency                                                              | ✅ Pinned in `clawql-api`                                                                                                                    |
+| `SearchService` / `ExecuteService`                                               | ✅ Live Layers; MCP uses `getClawqlApi().run(Effect…)`                                                                                       |
+| `AuditLive`                                                                      | ✅ Composed in `createClawQLApi()`                                                                                                           |
+| `Plugin` / `PluginRegistry`                                                      | ✅ Effect `register` / `beforeCallTool`                                                                                                      |
+| Extracted packages (`memory`, `documents`, `automation`, `sandbox`, `ouroboros`) | 🚧 Layer wrappers shipped; memory ingest/recall, documents IDP runner, + ouroboros hot paths use native `Effect.gen` (IO still `tryPromise`) |
+| `@effect/schema` at boundaries                                                   | ❌ Zod remains at MCP tool registration                                                                                                      |
+| Horizontal `Plugin` Layers                                                       | ✅ All tiers via `composeHorizontalPluginLayers()`                                                                                           |
+| Operator dynamic Layer list from CRD                                             | ✅ `composeHorizontalPluginLayersFromTierSpec()` maps `ClawQLHorizontalTierSpec` → Layers                                                    |
+| `MemoryIngestService` / vault post-sync                                          | ✅ Native `Effect.gen` stages prepare → vault write → `MemoryDbService` sync (no nested `runMemoryEffect`)                                   |
+| `DocumentsToolsService` IDP runner                                               | ✅ Native `Effect.gen` hop loop (skip/dry-run/execute+retry/Merkle/onHop); Promise façade kept for tests                                     |
 
 **Rule for new code in extracted packages:** prefer Effect in `clawql-core` / `clawql-api`; legacy `async` is acceptable at IO edges during migration (`Effect.tryPromise`). See plan §7.
